fix(storybook): single shared auth fetch stub so per-story admin flags actually apply
Storybook bundles every story module into one preview, so the four
per-file window.fetch shims chained and whichever installed last answered
/auth/me with its own module's flag — SettingsPage's default-true shim made
every AuthGate story render as admin regardless of the selected story.
Replaced with one shared installer + setStoryIsAdmin(), and keyed the
isAdmin decorators by story id so switching stories remounts AuthGate
instead of serving its cached /auth/me.
